Implementações de controle na pemrissão de visualização e edição com base nos relacionamentos do usuário

+++ HelpDesk.Business/Services/ClienteService.cs
@@ -1,4 +1,5 @@
-﻿using HelpDesk.Business.Interfaces.Repositories;
+﻿using HelpDesk.Business.Interfaces.Others;
+using HelpDesk.Business.Interfaces.Repositories;
 using HelpDesk.Business.Interfaces.Services;
 using HelpDesk.Business.Interfaces.Validators;
 using HelpDesk.Business.Models;
@@ -11,27 +12,67 @@ namespace HelpDesk.Business.Services
         private readonly IClienteRepository _clienteRepository;
         private readonly IClienteValidator _clienteValidator;
         private readonly IEnderecoRepository _enderecoRepository;
+        private readonly IUsuarioRepository _usuarioRepository;
+        private readonly IUser _user;
 
         public ClienteService(IClienteRepository clienteRepository,
                               IClienteValidator clienteValidator,
-                              IEnderecoRepository enderecoRepository)
+                              IEnderecoRepository enderecoRepository,
+                              IUsuarioRepository usuarioRepository,
+                              IUser user)
         {
             _clienteRepository = clienteRepository;
             _clienteValidator = clienteValidator;
             _enderecoRepository = enderecoRepository;
+            _usuarioRepository = usuarioRepository;
+            _user = user;
+        }
+        public async Task<IEnumerable<Cliente>> ObterTodos(int skip, int take)
+        {
+            var usuario = await _usuarioRepository.ObterUsuarioPorAutenticacao(_user.GetUserId());
+
+            var idGerenciadores = await _usuarioRepository.ObterGerenciadoresClientesPermitidos(usuario.Id);
+
+            return await _clienteRepository.ObterClientesPorPermissao(idGerenciadores.IdGerenciadores, skip, take);
+
         }
 
+        public async Task<Cliente?> ObterPorId(Guid id)
+        {
+            var usuario = await _usuarioRepository.ObterUsuarioPorAutenticacao(_user.GetUserId());
+
+            var idGerenciadoresUsuario = await _usuarioRepository.ObterGerenciadoresClientesPermitidos(usuario.Id);
+
+            var cliente = await _clienteRepository.ObterPorId(id);
+
+            if (_clienteValidator.ValidaPermissaoVisualizacao(cliente, idGerenciadoresUsuario.IdGerenciadores))
+            {
+                return cliente;
+            }
+
+            return null;
+        }
         public async Task Adicionar(Cliente cliente)
         {
+            var usuario = await _usuarioRepository.ObterUsuarioPorAutenticacao(_user.GetUserId());
+
+            var idGerenciadoresUsuario = await _usuarioRepository.ObterGerenciadoresClientesPermitidos(usuario.Id);
+
             if (await _clienteValidator.ValidaExistenciaPessoa(cliente.Id) 
-                || !await _clienteValidator.ValidaPessoa(new ClienteValidation(), cliente)) return;
+                || !await _clienteValidator.ValidaPessoa(new ClienteValidation(), cliente)
+                || !_clienteValidator.ValidaPermissaoInsercaoEdicao(cliente,idGerenciadoresUsuario.IdGerenciadores)) return;
 
             await _clienteRepository.Adicionar(cliente);
         }
 
         public async Task Atualizar(Cliente cliente)
         {
-            if (!await _clienteValidator.ValidaPessoa(new ClienteValidation(), cliente)) return;
+            var usuario = await _usuarioRepository.ObterUsuarioPorAutenticacao(_user.GetUserId());
+
+            var idGerenciadoresUsuario = await _usuarioRepository.ObterGerenciadoresClientesPermitidos(usuario.Id);
+
+            if (!await _clienteValidator.ValidaPessoa(new ClienteValidation(), cliente)
+                || !_clienteValidator.ValidaPermissaoInsercaoEdicao(cliente, idGerenciadoresUsuario.IdGerenciadores)) return;
 
             await _clienteRepository.Atualizar(cliente);
         }
@@ -45,7 +86,14 @@ namespace HelpDesk.Business.Services
 
         public async Task Remover(Guid idCliente)
         {
-            if (!await _clienteValidator.ValidaExclusaoCliente(idCliente)) return;
+            var usuario = await _usuarioRepository.ObterUsuarioPorAutenticacao(_user.GetUserId());
+
+            var idGerenciadoresUsuario = await _usuarioRepository.ObterGerenciadoresClientesPermitidos(usuario.Id);
+
+            var cliente = await _clienteRepository.ObterPorId(idCliente);
+
+            if (!await _clienteValidator.ValidaExclusaoCliente(idCliente)
+                || !_clienteValidator.ValidaPermissaoInsercaoEdicao(cliente, idGerenciadoresUsuario.IdGerenciadores)) return;
             
             await _clienteRepository.Remover(idCliente);
         }
